# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Synapse - An innovative language learning platform using AI-powered techniques\n\nFeatures:\n- 🎯 The Analyzer: Universal word analysis with definitions, conjugations, and examples\n- ✍️ The Scribe: AI-generated writing quests for practice\n- 🧠 The Synapse: Mind map visualization of learned words (Ghost → Liquid → Solid)\n- 🔍 The Lens: Import and analyze content from web articles\n- 🗣️ The Orator: Speaking coach with voice recognition\n- 📚 Spaced Repetition System (SRS): Scientific memorization using SM-2 algorithm",
        "title": "Synapse API",
        "contact": {
            "name": "Synapse Support",
            "url": "https://github.com/BachirKhiati/synapse/issues",
            "email": "support@synapse.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/{userID}/quests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all quests for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quests"
                ],
                "summary": "Get user quests",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of quests",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_BachirKhiati_synapse_internal_models.Quest"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{userID}/quests/generate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate a new AI-powered writing quest using user's ghost words",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quests"
                ],
                "summary": "Generate new quest",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated quest",
                        "schema": {
                            "$ref": "#/definitions/github_com_BachirKhiati_synapse_internal_models.Quest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Quest generation failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{userID}/quests/validate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validate user's written text for a quest using AI feedback",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quests"
                ],
                "summary": "Validate quest submission",
                "parameters": [
                    {
                        "description": "Quest validation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_BachirKhiati_synapse_internal_models.QuestValidationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Validation result with feedback",
                        "schema": {
                            "$ref": "#/definitions/github_com_BachirKhiati_synapse_internal_models.QuestValidationResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Quest not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Validation failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_BachirKhiati_synapse_internal_models.Quest": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "description": "beginner, intermediate, advanced",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "solution": {
                    "description": "The \"Glimpse Solution\"",
                    "type": "string"
                },
                "status": {
                    "description": "pending, in_progress, completed",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_BachirKhiati_synapse_internal_models.QuestValidationRequest": {
            "type": "object",
            "properties": {
                "quest_id": {
                    "type": "integer"
                },
                "user_text": {
                    "type": "string"
                }
            }
        },
        "github_com_BachirKhiati_synapse_internal_models.QuestValidationResponse": {
            "type": "object",
            "properties": {
                "feedback": {
                    "description": "Socratic guidance",
                    "type": "string"
                },
                "is_valid": {
                    "type": "boolean"
                },
                "new_words": {
                    "description": "Words used correctly for first time",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}